LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGAnimatedInteger-expected.txt

CONSOLE MESSAGE: Warning: feConvolveMatrix: problem parsing order="-1.5". Filtered element will not be displayed.

FAIL SVGAnimatedInteger interface - utilizing the targetX property of SVGFEConvolveMatrix assert_equals: expected 3 but got -1
